lfluxproject/lqa/requirements.txt
django-voting
django-threaded-comments

